+++ R/mod_tab_finds_aoristic.R
@@ -101,7 +101,7 @@ mod_aoristic_finds_serv <- function(id) {
       ns <- NS(id)
 
       generateCategorySelector("categories",
-                               parent = "Find",
+                               parent_category = "Find",
                                inputid = ns("selected_categories"))
 
       observeEvent(input$selected_categories, {